Refactor archive form into ArchiveSection component
Moved the poem submission form and related content from app/page.tsx into a new ArchiveSection component for better modularity and reuse. Updated the archive page layout and title, added Footer to the archive page, and removed the favicon.ico file.
